backend/requirements.txt
fastapi
uvicorn
python-multipart
praat-parselmouth
numpy
soundfile